CLI: Fix problem related to git parent repo
We are currently building our packages in a "packages" git ignored folder. In the case of swift, we need to operate a git repository. Having a git repo inside another git repo causes
problems, as sometimes we observed commands applied to
child repository were affecting the parent.

The solution: We are moving the build of the swift package
outside our working tree. We create a random /tmp subfolder with an uuid and leave the absolute path to it in the working tree ignored folder. In that way, developers are aware and machines can read that pointer to continue operating the package in further
steps.